test(session): hoist the shared paused-session and policy-override recipes (#559)
`createPausedSession` (create a MOCK session, launch it, run the timers,
simulate the entry stop, drop the launch's own DAP traffic) and the "spy on
the session store's `selectPolicy` and overlay hooks" recipe were each copied
per test file. Both now live in `session-manager-test-utils` —
`createPausedSession(sessionManager, dependencies, { clearDapCalls })` and
`overridePolicy(sessionManager, overrides)` — and `session-manager-dap.test.ts`
reads them from there instead of keeping its own.

A pure import swap: no assertion changed, all 120 tests in that file still
pass unchanged.
